# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP CUMBRE PETRÓLEO, GAS Y ENERGÍA",
        "description": "<img src='https://cumbrepetroleoygas.net/images/logo.svg' alt='Logo' width='30'>API oficial de la aplicación 'CUMBRE'<br><br><img src='https://codant.one/img/logos/codant.svg' alt='Logo' width='50'>Hecho por <a href='https://codant.one' target='_blank'>CODANT</a>",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.app-admin-cumbre/api",
            "description": "API"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "AUTH"
                ],
                "summary": "User login",
                "description": "Give a user access to the page",
                "operationId": "a5385ffca73f113abef00ea604d1ca58",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "fcm_token",
                                    "device_type",
                                    "lang"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "The user's E-mail",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User alphanumeric password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "fcm_token": {
                                        "description": "Firebase Token",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "device_type": {
                                        "description": "Device type (ios/android)",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "lang": {
                                        "description": "App language (es/en)",
                                        "type": "string",
                                        "format": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Some was wrong",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "AUTH"
                ],
                "summary": "Logout",
                "description": "Logout User",
                "operationId": "0f6715df0d22b69c7e35bd32b4c1fc92",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}